% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claim.R
\docType{data}
\name{Claim}
\alias{Claim}
\title{Claim}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Claim
}
\description{
Create a claim.
}
\examples{
\dontrun{
claim <- Claim$new() 
}

}
\keyword{datasets}
