% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{fact_check}
\alias{fact_check}
\title{Check}
\usage{
fact_check(q = NULL, lang = NULL, publisher = NULL,
  days_old = NULL, pages = 1)
}
\arguments{
\item{q}{Textual query string. Required unless \code{publisheer} is specified.}

\item{lang}{The BCP-47 language code, such as \code{en-US} or \code{sr-Latn}. 
Can be used to restrict results by language, though we do not currently consider the region.}

\item{publisher}{The review publisher site to filter results by, e.g. \code{nytimes.com}.}

\item{days_old}{The maximum age of the returned search results, in days. 
Age is determined by either claim date or review date, whichever is newer.}

\item{pages}{Number of pages of results to fecth.}
}
\description{
Fact Check publishers or queries.
}
\examples{
\dontrun{
# Fact checks and claims on Trump only in English
fact_check("Trump", lang = "en-US")

# fact checks of the New York Times
fact_check(publisher = "nytimes.com")
}

}
