% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{factcheck_key}
\alias{factcheck_key}
\title{Setup}
\usage{
factcheck_key(key)
}
\arguments{
\item{key}{Your API key, freely available at \url{https://console.cloud.google.com}.}
}
\description{
Setup your session, all subsequent calls will be done using the API key.
}
\note{
You can specify \code{FACTCHECK_API_KEY} as environment variable, likely in your \code{.Renviron} file.
}
