% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth.R
\name{fact_token}
\alias{fact_token}
\title{OAuth}
\usage{
fact_token(cache = gargle::gargle_oauth_cache(),
  use_oob = gargle::gargle_oob_default())
}
\arguments{
\item{cache}{Whether to cache the token.}

\item{use_oob}{Whether to use out-of-band authentication.}
}
\description{
Create an OAuth token.
}
