% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth.R
\name{fact_token}
\alias{fact_token}
\title{OAuth}
\usage{
fact_token(token = Sys.getenv("FACTCHECK_TOKEN"), id, secret)
}
\arguments{
\item{token}{A previously generated token or the path to the saved \code{RDS} file.}

\item{id, secret}{Your credentials.}
}
\value{
Invisibly returns an object of class \code{Token2.0}.
}
\description{
Create an OAuth token.
}
\examples{
\dontrun{
token <- fact_token(id = "XxxXXXxxxX", secret = "XXxxXxX")

# In future sessions pass Token  object
fact_token(token)
}

}
